Move rabbitListen to RabbitMQ.ts
