+++ src/util/util/ipc/RabbitMQ.ts
@@ -16,9 +16,12 @@
 	along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
+import { randomUUID } from "node:crypto";
 import EventEmitter from "node:events";
 import amqp, { Channel, ChannelModel } from "amqplib";
+import { EVENT } from "../../interfaces";
 import { Config } from "../Config";
+import type { EventOpts } from "./Event";
 
 export class RabbitMQ {
     public static connection: ChannelModel | null = null;
@@ -170,4 +173,52 @@ export class RabbitMQ {
     static isConnected(): boolean {
         return this.connection !== null && !this.isReconnecting;
     }
+}
+
+export async function rabbitListen(channel: Channel, id: string, callback: (event: EventOpts) => unknown, opts?: { acknowledge?: boolean }): Promise<() => Promise<void>> {
+    await channel.assertExchange(id, "fanout", { durable: false });
+    const q = await channel.assertQueue("", {
+        exclusive: true,
+        autoDelete: true,
+        messageTtl: 5000,
+    });
+
+    const consumerTag = randomUUID();
+
+    const cancel = async () => {
+        try {
+            await channel.unbindQueue(q.queue, id, "");
+            await channel.cancel(consumerTag);
+        } catch (e) {
+            console.log("[RabbitMQ] Error while cancelling channel (may be expected):", e instanceof Error ? e.message : e);
+        }
+    };
+
+    await channel.bindQueue(q.queue, id, "");
+    await channel.consume(
+        q.queue,
+        (opts) => {
+            if (!opts) return;
+
+            const data = JSON.parse(opts.content.toString());
+            const event = opts.properties.type as EVENT;
+
+            callback({
+                event,
+                data,
+                acknowledge() {
+                    channel.ack(opts);
+                },
+                channel,
+                cancel,
+            });
+            // rabbitCh.ack(opts);
+        },
+        {
+            noAck: !opts?.acknowledge,
+            consumerTag: consumerTag,
+        },
+    );
+
+    return cancel;
 }
